feat(router,app): allows access with REST-like URLs to statistics page
Now users can use routes such as `/results/stage/main_4/main_04-03` to access the statistics in zone
`main_4` stage `main_04-03`. Such URL gives flexability to share the page with others that, the
receiver of such URL will be deeplinked to the actural resource they were looking at, instead of
asking them again.

BREAKING CHANGE: The name of components were now depends on how many information the user provides
in the URL. If there's any navigation or redirection towards such resource it must be edited to suit
the new router structure. Although the old component name stays as-is and the code before such
change should still redirect the user to the index page, it is still good to check that if any
unexpected behavior occurs.