improvement(vuetify): migrated to vuetify v2.2.4
migrated to vuetify v2.2.4 by following the official documentation at
https://vuetifyjs.com/zh-Hans/getting-started/releases-and-migrations

BREAKING CHANGE: In this commit, StageSelector has been deprecated and switched to the new
NewStageSelector component, but the new component is still under construction. This commit is to act
as a milestone of migrating vuetify to the new version in order to improve clearity. Also, the page
layout at `Home`, `Members` and `Donate` are modified due to the vuetify migration. This commit
should NOT be merged until the StageSelector has been completely refactored.

#73