refactor(preloader): rewrite preloader mechanism to avoid using gif
mobile safari does not support transparent gif - or at least it behaves in an unexpected way which -
it displays white background on some of the frames instead of transparent background (such gif works
well on chrome). in order to avoid poor UX we have migrated the solution to be using a PNG sequence
in order to compensate such lack of feature that mobile safari has.